Added a User schema with authentication and JWT support
Implemented a Mongoose schema to manage all the users for the platform. The schema has fields for email, password and role. Add bycrypt to encrypt the password before saving. Added method to compare passwords with the JWT for user authentication. Added environment variables for JWT_SECRET, JWT_EXPIRE, and JWT_COOKIE_EXPIRE.